Split locomotion environments.
